requirements-web.txt
# Web application dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.13.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
websockets==12.0
pandas==2.1.4
plotly==5.17.0
kaleido==0.2.1

# Keep existing scientific dependencies
PyYAML>=6.0
oemof.solph>=0.5.0
matplotlib>=3.5.0
numpy>=1.21.0
